requirements.txt
Flask==2.0.2
Flask-RESTful==0.3.9
mysql-connector==2.2.9
mysql-connector-python==8.0.17
pandas==1.3.4
setuptools-scm==6.3.2